Move WykopWebViewClient to separate file
