+++ app/src/main/kotlin/io/github/feelfreelinux/wykopmobilny/activities/LoginActivity.kt
@@ -1,31 +1,32 @@
 package io.github.feelfreelinux.wykopmobilny.activities
 
-import android.annotation.TargetApi
-import android.os.Build
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
-import android.webkit.WebResourceRequest
-import android.webkit.WebView
-import android.webkit.WebViewClient
 import io.github.feelfreelinux.wykopmobilny.R
 import io.github.feelfreelinux.wykopmobilny.objects.APP_KEY
 import io.github.feelfreelinux.wykopmobilny.utils.ApiPreferences
 import io.github.feelfreelinux.wykopmobilny.utils.WykopApiManager
+import io.github.feelfreelinux.wykopmobilny.utils.invisible
 import kotlinx.android.synthetic.main.activity_login.*
 
 class LoginActivity : AppCompatActivity() {
 
     private val apiPreferences = ApiPreferences()
-    private val apiManager: WykopApiManager by lazy {
-        WykopApiManager(this)
-    }
+    private val webViewClient = WykopWebViewClient()
+    private val apiManager: WykopApiManager by lazy { WykopApiManager(this) }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_login)
         title = getString(R.string.login)
-
+        setupLoginViewClient()
         checkIfUserIsLogged()
+    }
+
+    private fun setupLoginViewClient() {
+        webViewClient.onLoginSuccess {
+            startMikroblog()
+        }
     }
 
     private fun checkIfUserIsLogged() {
@@ -37,45 +38,13 @@ class LoginActivity : AppCompatActivity() {
     }
 
     private fun showLoginWebView() {
-        // Set custom url loader to WebView
-        webView.webViewClient = object : WebViewClient() {
-            @SuppressWarnings("deprecation")
-            override fun shouldOverrideUrlLoading(view: WebView?, url: String?): Boolean {
-                url?.let { parseUrl(it) }
-                return false
-            }
-
-            @TargetApi(Build.VERSION_CODES.LOLLIPOP)
-            override fun shouldOverrideUrlLoading(view: WebView?, request: WebResourceRequest?): Boolean {
-                request?.let { parseUrl(it.url.toString()) }
-                return false
-            }
-
-            private fun parseUrl(redirectUrl: String) {
-                if (redirectUrl.containsUserData()) {
-                    saveDataToPrefs(redirectUrl)
-                    startMikroblog()
-                }
-            }
-
-            private fun saveDataToPrefs(redirectUrl: String) {
-                apiPreferences.login = redirectUrl
-                        .split("/token/").first()
-                        .substringAfterLast("/login/")
-
-                apiPreferences.userKey = redirectUrl
-                        .split("/token/").last()
-                        .replace("/", "")
-            }
-
-            private fun String.containsUserData() = contains("/token/") and contains("/login/")
-        }
-        // Open wykop connect api site
+        webView.webViewClient = webViewClient
         webView.loadUrl("http://a.wykop.pl/user/connect/appkey/$APP_KEY")
     }
 
     fun startMikroblog() {
         supportActionBar?.hide()
+        webView.invisible()
         apiManager.getUserSessionToken(
                 successCallback = {
                     launchMikroblogHotList(apiManager.getData())
@@ -83,3 +52,4 @@ class LoginActivity : AppCompatActivity() {
                 })
     }
 }
+
